reorganize frontend folders & files
