+++ package.json
@@ -1,44 +1,38 @@
 {
-  "name": "fe",
+  "name": "GuarderiaFE",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "test:unit": "vue-cli-service test:unit",
+    "start": "node server.js"
   },
   "dependencies": {
+    "@apollo/client": "^3.4.17",
+    "@vue/apollo-composable": "^4.0.0-alpha.15",
+    "@vue/apollo-option": "^4.0.0-alpha.15",
+    "apollo-link-context": "^1.0.20",
     "core-js": "^3.6.5",
+    "express": "^4.17.1",
+    "graphql": "^16.0.1",
+    "graphql-tag": "^2.12.6",
+    "jwt-decode": "^3.1.2",
+    "package.json": "^2.0.1",
+    "serve-static": "^1.14.1",
     "vue": "^3.0.0",
-    "vue-router": "^4.0.0-0"
+    "vue-router": "^4.0.0-0",
+    "vuex": "^3.6.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-router": "~4.5.0",
+    "@vue/cli-plugin-unit-jest": "^4.5.13",
     "@vue/cli-service": "~4.5.0",
     "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-standard": "^5.1.2",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^7.0.0"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/vue3-essential",
-      "@vue/standard"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
+    "@vue/test-utils": "^2.0.0-rc.15",
+    "typescript": "~3.9.3",
+    "vue-jest": "^5.0.0-0"
   },
   "browserslist": [
     "> 1%",
