+++ src/components/About.vue
@@ -1,0 +1,22 @@
+<template>
+  <div class="about">
+    <h1>This is an about page</h1>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'AboutUs',
+  setup() {
+    
+  },
+}
+</script>
+
+<style>
+
+.about h1{
+  justify-content: center;
+}
+
+</style>
