+++ src/main.js
@@ -1,5 +1,31 @@
 import { createApp } from 'vue'
+import { ApolloClient, createHttpLink, InMemoryCache } from '@apollo/client/core'
+import { createApolloProvider } from '@vue/apollo-option'
+import { setContext } from 'apollo-link-context'
 import App from './App.vue'
 import router from './router'
 
-createApp(App).use(router).mount('#app')
+
+const httpLink = createHttpLink({
+    uri: 'https://mintic-p5-g4-dw-apigateway.herokuapp.com/',
+})
+
+const authLink = setContext((_, { headers }) => {
+    return {
+        headers: {
+            ...headers,
+            "Authorization": localStorage.getItem("token_acces") || ""
+        }
+    }
+})
+
+const apolloClient = new ApolloClient({
+    link: authLink.concat(httpLink),
+    cache: new InMemoryCache()
+})
+
+const apolloProvider = new createApolloProvider({
+    defaultClient: apolloClient
+})
+
+createApp(App).use(router).use(apolloProvider).mount('#app')
