--- src/router/index.js
@@ -1,39 +1,0 @@
-import { createRouter, createWebHashHistory } from 'vue-router'
-import Home from '../views/Home.vue'
-import LogIn from '../views/LogIn.vue'
-import products from '../views/Products.vue'
-
-const routes = [
-  
-  {
-  
-    path: '/',
-    name: 'Home',
-    component: Home
-  },
-  {
-    path: '/about',
-    name: 'About',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-  },
-  {
-    path: '/logIn',
-    name: "logIn",
-    component: LogIn 
-},
-  {
-    path: '/products',
-    name: "products",
-    component: products 
-  }
-]
-
-const router = createRouter({
-  history: createWebHashHistory(),
-  routes
-})
-
-export default router
